further fix

+++ 0x00-python_variable_annotations/7-to_kv.py
@@ -7,4 +7,4 @@ from typing import Tuple, Union
 
 def to_kv(k: str, v: Tuple[Union[int, float]]) -> Tuple[str, float]:
     """Crazy function with no purpose except learning annotation"""
-    return k, v*v
+    return k, float(v*v)
